requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-rest-auth==0.9.5
djangorestframework==3.10.2
idna==2.8
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
